Split resolve tests into dedicated module
